+++ src/modules/python.rs
@@ -1,6 +1,6 @@
 use super::Segment;
 use crate::context::Context;
-use crate::utils;
+use crate::utils::project_detector;
 use ansi_term::Color;
 use std::process::Command;
 
@@ -14,11 +14,11 @@ use super::{Context, Module};
 ///     - Current directory contains a `requirements.txt` file
 ///     - Current directory contains a `pyproject.toml` file
 pub fn segment(context: &Context) -> Option<Segment> {
-    let python_criteria = utils::Criteria::new()
+    let python_criteria = project_detector::Criteria::new()
         .set_files(vec!["requirements.txt, pyproject.toml", ".pyproject.toml"])
         .set_extensions(vec!["py"]);
 
-    let is_py_project = utils::is_lang_project(&context.dir_files, &python_criteria);;
+    let is_py_project = project_detector::is_lang_project(&context.dir_files, &python_criteria);;
     if !is_py_project {
         return None;
     }
