+++ src/modules/rust.rs
@@ -1,6 +1,6 @@
 use super::Segment;
 use crate::context::Context;
-use crate::utils;
+use crate::utils::project_detector;
 use ansi_term::Color;
 use std::process::Command;
 
@@ -12,11 +12,11 @@ use super::{Context, Module};
 ///     - Current directory contains a file with a `.rs` extension
 ///     - Current directory contains a `Cargo.toml` file
 pub fn segment(context: &Context) -> Option<Segment> {
-    let rust_criteria = utils::Criteria::new()
+    let rust_criteria = project_detector::Criteria::new()
         .set_files(vec!["Cargo.toml"])
         .set_extensions(vec!["rs"]);
 
-    let is_rs_project = utils::is_lang_project(&context.dir_files, &rust_criteria);
+    let is_rs_project = project_detector::is_lang_project(&context.dir_files, &rust_criteria);
     if !is_rs_project {
         return None;
     }
