+++ src/utils/project_detector.rs
